requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-celery-beat==2.2.1
django-environ==0.4.5
django-redis==5.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
